Lisp: Added * and or ops to lisp_type_spec_matches
These allow, for example, the expression (* (or f32 i32)), which
represents "any number of f32 or i32", which is what the *, - and +
primitives accept.  We were only able to simplify the implementation
of prim_concat, but 'or' will also be useful for ECS primitives that
have to accept an Entity or a Relationship, like ecs-add.