Entity names & stdlib.lisp
Entities can now have names, in the form of Lisp symbols.  An Entity's
name is unique within the World, and is *not* a Component.  The main
use of this feature is to look up Entities by name, which makes it
easier to access pre-created Entities, especially standard Components
like Pos.

We now load all "standard" Lisp files in stdlib.lisp, rather than
loading each of them from within C.