include/lisp/lisp.h: LISP_ASSERT lisp_type_name
The LISP_ASSERT_RAW_TYPE macro uses OBJ_TYPE to test the type of the
argument, so it is not expected to distinguish between struct types,
so it is more correct (and efficient) to us lisp_type_name instead of
lisp_type_of.

The fact this got through highlights how poor of a naming convention
this is.